start a new video every midnight
When a TCP socket opens up and video data start pouring in trough it,
create a folder with it's name in YYYY-MM format if it doesn't exist
yet, and create a new file in it with the format DD_HH:mm:ss.h264.

If the camera gets disconnected or it's 00:00, close the file and rename
it DD_HH:mm:ss_-_DD_HH:mm:ss.h264, where the first part is the start of
recording, and the second part is the time when the file was closed.
After this, create a new file for the stream.